requirements.txt
numpy
pandas
scipy
scikit-learn
matplotlib
tqdm
pyyaml
pydub
soundfile
librosa
torch
torchvision
torchaudio
webrtcvad
noisereduce
opensmile
shap
tensorboard
seaborn
captum
transformers
peft